Move queryClient instantiation to utils/query.ts
