+++ apps/client/src/utils/query.ts
@@ -1,0 +1,8 @@
+import { QueryClient } from "@tanstack/react-query"
+
+export const queryClient = new QueryClient({
+    defaultOptions: {
+        queries: {},
+        mutations: {}
+    }
+})
